+++ src/main/java/com/comeon/backend/user/command/application/UserFacade.java
@@ -1,0 +1,21 @@
+package com.comeon.backend.user.command.application;
+
+import com.comeon.backend.user.command.domain.User;
+import com.comeon.backend.user.command.domain.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class UserFacade {
+
+    private final UserRepository userRepository;
+
+    public void modifyUser(Long userId, String nickname, String profileImageUrl) {
+        User user = userRepository.findBy(userId).orElseThrow();
+        user.updateNickname(nickname);
+        user.updateProfileImage(profileImageUrl);
+    }
+}
