+++ src/main/java/com/comeon/backend/user/command/domain/GoogleOauthService.java
@@ -1,0 +1,6 @@
+package com.comeon.backend.user.command.domain;
+
+public interface GoogleOauthService {
+
+    OauthUserInfo getUserInfoByIdToken(String idToken);
+}
