+++ src/main/java/com/comeon/backend/user/command/domain/KakaoOauthService.java
@@ -1,0 +1,6 @@
+package com.comeon.backend.user.command.domain;
+
+public interface KakaoOauthService {
+
+    OauthUserInfo getUserInfoByCode(String code);
+}
