+++ src/main/java/com/comeon/backend/user/command/infra/repository/UserJpaRepository.java
@@ -1,0 +1,14 @@
+package com.comeon.backend.user.command.infra.repository;
+
+import com.comeon.backend.user.command.domain.OauthProvider;
+import com.comeon.backend.user.command.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserJpaRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByOauthIdAndProvider(String oauthId, OauthProvider provider);
+    List<User> findByIdIn(List<Long> userIds);
+}
