--- src/main/java/com/comeon/backend/user/infrastructure/feign/google/GoogleUserInfoResponse.java
@@ -1,13 +1,0 @@
-package com.comeon.backend.user.infrastructure.feign.google;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GoogleUserInfoResponse {
-
-    private String oauthId;
-    private String email;
-    private String name;
-}
