--- src/main/java/com/comeon/backend/user/presentation/api/request/KakaoOAuth2LoginRequest.java
@@ -1,16 +1,0 @@
-package com.comeon.backend.user.presentation.api.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.validation.constraints.NotBlank;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class KakaoOAuth2LoginRequest {
-
-    @NotBlank
-    private String code;
-}
