--- src/main/java/com/comeon/backend/user/presentation/api/request/UserModifyRequest.java
@@ -1,14 +1,0 @@
-package com.comeon.backend.user.presentation.api.request;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserModifyRequest {
-
-    private String nickname;
-    private String profileImageUrl;
-}
