--- src/main/java/com/comeon/backend/user/presentation/api/response/UserModifyResponse.java
@@ -1,13 +1,0 @@
-package com.comeon.backend.user.presentation.api.response;
-
-import lombok.Getter;
-
-@Getter
-public class UserModifyResponse {
-
-    private Boolean success;
-
-    public UserModifyResponse() {
-        this.success = true;
-    }
-}
