+++ src/main/java/com/comeon/backend/user/presentation/request/KakaoOauthRequest.java
@@ -1,0 +1,16 @@
+package com.comeon.backend.user.presentation.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.validation.constraints.NotBlank;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class KakaoOauthRequest {
+
+    @NotBlank
+    private String code;
+}
