+++ src/main/java/com/comeon/backend/user/presentation/response/UserDetailsResponse.java
@@ -1,0 +1,16 @@
+package com.comeon.backend.user.presentation.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserDetailsResponse {
+
+    private Long userId;
+    private String nickname;
+    private String profileImageUrl;
+    private String role;
+    private String email;
+    private String name;
+}
