+++ src/main/java/com/comeon/backend/user/presentation/response/UserModifyResponse.java
@@ -1,0 +1,13 @@
+package com.comeon.backend.user.presentation.response;
+
+import lombok.Getter;
+
+@Getter
+public class UserModifyResponse {
+
+    private Boolean success;
+
+    public UserModifyResponse() {
+        this.success = true;
+    }
+}
